requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
heroku==0.1.4
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
python-dateutil==1.5
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
whitenoise==4.1.2
gunicorn==19.9.0
django-cors-headers
